refactor test fixtures into seperate file and extract helper methods into _test_utils
